### 2025-07-15 #8588 @editorialbot
[PRE REVIEW]: caskade: building Pythonic scientific simulators
**Submitting author:** @ConnorStoneAstro (<a href="http://orcid.org/0000-0002-9086-6398">Connor Stone</a>)
**Repository:** https://github.com/ConnorStoneAstro/caskade
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.10.3
