### 2025-07-15 #8588 @danielskatz
👋 @ConnorStoneAstro - thanks for your submission. Please either add a Statement of Need section to your paper, or rename a current section to that.

In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
