### 2025-07-16 6c14efb
use log scale for single valid, use clamp for full valid
