### 2025-07-17 90ef0d9
add statement of need to paper, and missing doi
