+++ paper.md
@@ -53,79 +53,85 @@ bibliography: paper.bib
 Scientific simulators and pipelines form the core of many research projects.
 Writing high quality, modular code allows for efficiently scaling a project, but
 this can be challenging in a research context. Research project goals and
-solutions to those goals are constantly in flux, often requiring many
-refactoring rounds to meet these changes. The result can be a progressively more
-unwieldy interconnected code; we dub this, "the args problem" in scientific
-simulator construction. Here we present a system, `caskade`, which allows users
-to focus on modular components of a simulator, these are small and testable to
+solutions to those goals are constantly in flux, requiring many refactoring
+rounds to meet these changes. The result can be a progressively more unwieldy
+interconnected code. Here we present a system, `caskade`, which allows users to
+focus on modular components of a simulator, these are small and testable to
 ensure robustness. With `caskade` one can turn these modular components into
-abstracted blocks that stack to form a powerful simulator. This is inspired by
+abstracted blocks that connect to form a powerful simulator. `caskade` manages
+the flow of parameter values through such a simulator. 
+
+# Statement of Need
+
+Science is an intrinsically iterative process, and so is the development of
+scientific code. Well written code is flexible and scalable while being
+performant, this is difficult to achieve in a scientific context where goals
+often evolve rapidly, requiring code refactoring. A major aspect of this is the
+parameters of a scientific model, the values that will ultimately be sampled
+and/or optimized to represent some data. A value may need to alternately be
+fixed, then allowed to vary (e.g. in Gibbs sampling). Some parameters that were
+initially separate may need to share a value or some functional relationship. In
+the extreme, a whole simulator may become a function of a single variable, such
+as time. Meta-data such as the uncertainty or valid range of a parameter may
+need to be stored. One may need to represent all parameters as a single 1D
+vector to interface with external tools, such as emcee [@emcee], scipy.optimize
+[@scipy], Pyro [@pyro], dynesty [@dynesty], and torch.Optim [@pytorch]. Large
+projects and correspondingly large teams require the ability to break projects
+into manageable subtasks which can later be naturally combined into a complete
+analysis suite. Most importantly, as all of the above needs change, it is
+critical to meaningfully re-use older code without "code debt" or "software
+entropy" growing unsustainably.
+
+# Features
+
+The core features of `caskade` are the `Module` base class, `Param` object, and
+`forward` decorator. To construct a `caskade` simulator, one subclasses `Module`
+then adds some number of `Param` objects as attributes of the class. Any number
+of class methods may be decorated with `@forward`, meaning `caskade` will manage
+the `Param` arguments of that function. As modules are combined into a larger
+simulator, `caskade` builds a directed acyclic graph (DAG) representation. This
+allows it to automatically manage the flow (cascade) of parameters through the
+simulator and encode arbitrary relationships between them. This is inspired by
 the `PyTorch` framework `nn.Module` which allows for near effortless
 construction of machine learning models. We generalize the object oriented
 framework to apply to almost any scientific forward model, simulator, analysis
 pipeline, and so on; `caskade` manages the flow of parameters through these
-models.
-
-# Features
-
-The core features of `caskade` are the `Module` base class, `Param` parameter,
-and `forward` decorator. To construct a `caskade` simulator, one subclasses
-`Module` then adds some number of `Param` objects as attributes of the class,
-finally any number of class methods may be decorated with `@forward`. As modules
-are combined into a larger simulator, `caskade` builds a directed acyclic graph
-(DAG) representation. This allows it to automatically manage the flow (cascade)
-of parameters through the simulator and encode arbitrary relationships between
[…]
