+++ paper.bib
@@ -35,6 +35,7 @@ author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and B
 booktitle = {Advances in Neural Information Processing Systems 32},
 pages = {8024--8035},
 year = {2019},
+doi = {10.48550/arXiv.1912.01703},
 publisher = {Curran Associates, Inc.},
 url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
@@ -96,6 +97,7 @@ archivePrefix = {arXiv},
   volume={20},
   number={1},
   pages={973--978},
+  doi = {10.48550/arXiv.1810.09538},
   year={2019},
   publisher={JMLR. org}
 }
@@ -129,6 +131,7 @@ archivePrefix = {arXiv},
   journal={Graph drawing software},
   pages={127--148},
   year={2004},
+  doi={10.1007/978-3-642-18638-7_6},
   publisher={Springer}
 }
 
