### 2025-07-17 #8588 @ConnorStoneAstro
@danielskatz I have rearranged the text to have a clear statement of need section. I have also added DOIs where possible. The JAX and Hydra packages provide clear instructions on how to cite them, which don't include DOIs. This is true for PyTorch and Pyro as well, though I was able to find an appropriate DOI for their arXiv versions. Please let me know if anything else needs updating.
