### 2025-07-26 b4accd1
fix node str bug
