### 2025-07-26 ee308ee
revert sigmoid keep log
