+++ src/caskade/param.py
@@ -543,6 +543,9 @@ class Param(Node):
         return value
 
     def _to_valid_fullvalid(self, value: ArrayLike) -> ArrayLike:
+        value = (
+            backend.logit((value - self.valid[0]) / (self.valid[1] - self.valid[0])) + self.valid[0]
+        )
         return value
 
     def _to_valid_cyclic(self, value: ArrayLike) -> ArrayLike:
@@ -558,7 +561,9 @@ class Param(Node):
         return value
 
     def _from_valid_fullvalid(self, value: ArrayLike) -> ArrayLike:
-        value = backend.clip(value, min=self.valid[0], max=self.valid[1])
+        value = (
+            backend.sigmoid(value - self.valid[0]) * (self.valid[1] - self.valid[0]) + self.valid[0]
+        )
         return value
 
     def _from_valid_cyclic(self, value: ArrayLike) -> ArrayLike:
