### 2025-07-26 8ab071e
use log scale for single valid, use clamp for full valid (#56)
* use log scale for single valid, use clamp for full valid

* revert sigmoid keep log
