+++ src/caskade/param.py
@@ -552,10 +552,10 @@ class Param(Node):
         return ((value - self.valid[0]) % (self.valid[1] - self.valid[0])) + self.valid[0]
 
     def _to_valid_leftvalid(self, value: ArrayLike) -> ArrayLike:
-        return value - 1.0 / (value - self.valid[0])
+        return backend.log(value - self.valid[0])
 
     def _to_valid_rightvalid(self, value: ArrayLike) -> ArrayLike:
-        return value + 1.0 / (self.valid[1] - value)
+        return backend.log(self.valid[1] - value)
 
     def _from_valid_base(self, value: ArrayLike) -> ArrayLike:
         return value
@@ -571,11 +571,11 @@ class Param(Node):
         return value
 
     def _from_valid_leftvalid(self, value: ArrayLike) -> ArrayLike:
-        value = (value + self.valid[0] + backend.sqrt((value - self.valid[0]) ** 2 + 4)) / 2
+        value = backend.exp(value) + self.valid[0]
         return value
 
     def _from_valid_rightvalid(self, value: ArrayLike) -> ArrayLike:
-        value = (value + self.valid[1] - backend.sqrt((value - self.valid[1]) ** 2 + 4)) / 2
+        value = self.valid[1] - backend.exp(value)
         return value
 
     @property
