### 2025-07-29 #8588 @drvinceknight
@ConnorStoneAstro if you could, do you mind taking a look at [this list of people](https://reviewers.joss.theoj.org/lookup) to suggest potential reviewers.
