### 2025-08-01 83db902
add description parameter for all node objects (#58)
* add description for node

* remove unused backend function
