### 2025-08-04 #8588 @drvinceknight
@yasahi-hpc @aslan-ng @Jashcraf @Christovis @avapolzin I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “caskade: building Pythonic scientific simulators”. You can find more information at the top of this Github issue.
