### 2025-08-13 #8786 @editorialbot
[REVIEW]: caskade: building Pythonic scientific simulators
**Submitting author:** @ConnorStoneAstro (<a href="http://orcid.org/0000-0002-9086-6398">Connor Stone</a>)
**Repository:** https://github.com/ConnorStoneAstro/caskade
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.11.0
