### 2025-08-14 8610e91
Add active state cache decorator
