+++ paper.bib
@@ -142,3 +142,52 @@ archivePrefix = {arXiv},
   year =         {2019},
   url =          {https://github.com/facebookresearch/hydra}
 }
+
+@article{Hafner2025, 
+doi = {10.21105/joss.08385}, 
+url = {https://doi.org/10.21105/joss.08385}, 
+year = {2025}, 
+publisher = {The Open Journal}, 
+volume = {10}, 
+number = {111}, 
+pages = {8385}, 
+author = {Häfner, Dion and Lavin, Alexander}, 
+title = {Tesseract Core: Universal, autodiff-native software components for Simulation Intelligence}, 
+journal = {Journal of Open Source Software} 
+} 
+
+@article{Stammler2022, 
+doi = {10.21105/joss.03882}, 
+url = {https://doi.org/10.21105/joss.03882}, 
+year = {2022}, 
+publisher = {The Open Journal}, 
+volume = {7}, 
+number = {69}, 
+pages = {3882}, 
+author = {Stammler, Sebastian M. and Birnstiel, Tilman}, 
+title = {Simframe: A Python Framework for Scientific Simulations}, 
+journal = {Journal of Open Source Software} 
+} 
+
+@article{Blochwitz2012,
+author = {Blochwitz, T., Otter, M., Åkesson, J., Arnold, M., Clauss, C., Elmqvist, H., Friedrich, M.,
+Junghanns, A., Mauss, J., Neumerkel, D., & others}, 
+title = {Functional mockup interface 2.0: The standard for tool independent exchange of simulation models}, 
+publisher = {9th International Modelica Conference}, 
+pages = {173-184}, 
+year = {2012}, 
+doi = {10.3384/ecp12076173}
+}
+
+@article{Hatledal2025, 
+doi = {10.21105/joss.08182}, 
+url = {https://doi.org/10.21105/joss.08182}, 
+year = {2025}, 
+publisher = {The Open Journal}, 
+volume = {10}, 
+number = {110}, 
+pages = {8182}, 
+author = {Hatledal, Lars Ivar}, 
+title = {Ecos: An accessible and intuitive co-simulation framework}, 
+journal = {Journal of Open Source Software} 
+} 
