+++ paper.md
@@ -96,9 +96,7 @@ the `PyTorch` framework `nn.Module` which allows for near effortless
 construction of machine learning models. We generalize the object oriented
 framework to apply to almost any scientific forward model, simulator, analysis
 pipeline, and so on; `caskade` manages the flow of parameters through these
-models. In some ways `caskade` is reminiscent of `Hydra` [@hydra], however
-`caskade` focuses on numerical parameters and scientific inference. The two may
-even be used in tandem.
+models.
 
 Thus the primary capability of `caskade` is the management of `Param` values as
 they enter `@forward` methods of `Modules`. Any parameter may be transformed
@@ -126,12 +124,30 @@ See also that all of the lens objects (`ExternalShear`, `SIE`, and
 `SinglePlane`) point to a single cosmology `Module` and so share the same
 cosmological parameters automatically.
 
-The `caskade` package maintains 100% coverage for unit testing. Further, we
-include tutorial notebooks which demonstrate all of the available features. The
-`caustics` package can serve as a project template utilizing the many features
-of `caskade`.
-
 [^1]: visual generated by `graphviz` [@graphviz]
+
+# State of the Field
+
+In some ways `caskade` is reminiscent of `Hydra` [@hydra], however `caskade`
+focuses on numerical parameters and scientific inference, while `Hydra` focuses
+on configuration management and large scale process organization. The two may
+even be used in tandem. Another package, `tesseract-core` [@Hafner2025] focuses
+more on containerization and distribution of simulations to interface different
+ecosystems (`PyTorch` and `JAX` as well as `Python` and `C++`) and on different
+compute engines (HPC clusters or in cloud). The `SimFrame` [@Stammler2022]
+package shares `caskade`'s modular and extensible core design, though is focused
+exclusively on solving differential equations. Encoding the Functional Mockup
+Interface standard [@Blochwitz2012] is the `Ecos` package [@Hatledal2025] which
+is also designed for building modular simulators though in the more strict FMI
+standard which requires auxiliary `.xml` specification files, `caskade` focuses
+on lean and active research development which thrives on minimal overhead.
+Finally, `PathSim` also shares the `caskade` modular simulator building
+framework, though it focuses exclusively on time-domain dynamical systems.
+Clearly, many fields of research and development desire such modular
+simulation-building frameworks; `caskade` fulfills the role very generally,
+though not so abstractly as to require overhead schema or meta-data files. It is
+lean and extensible so as to grow with a project.
+
 
 # Acknowledgements
 
