### 2025-08-19 43c91e3
minor changes cleanup presentation
