+++ paper.md
@@ -101,14 +101,12 @@ models.
 Thus the primary capability of `caskade` is the management of `Param` values as
 they enter `@forward` methods of `Modules`. Any parameter may be transformed
 between "static" and "dynamic" where static has a fixed value and dynamic is
-provided at call time. Individual parameters or whole branches of the DAG may be
-switched between static and dynamic. Parameters may be synced with arbitrary
-functional relationships between them. New parameters may be added dynamically
-to allow for sophisticated transformations. For example, an entire simulator may
-be turned into a function of time without modifying the underlying simulator by
-adding a time parameter and linking appropriately. It is possible to use
-`caskade` with `NumPy` [@numpy], `JAX` [@jax], or `PyTorch` [@pytorch] numerical
-backends.
+provided at call time. Parameters may be synced with arbitrary functional
+relationships between them. New parameters may be added dynamically to allow for
+sophisticated transformations. For example, an entire simulator may be turned
+into a function of time without modifying the underlying simulator by adding a
+time parameter and linking appropriately. It is possible to use `caskade` with
+`NumPy` [@numpy], `JAX` [@jax], or `PyTorch` [@pytorch] numerical backends.
 
 ![Example `caskade` DAG representation of a gravitational lensing simulator. Ovals represent Modules, boxes represent dynamic parameters, shaded boxes represent fixed parameters, arrow boxes represent parameters which are functionally dependent on another parameter, and thin arrows show the direction of the graph flow for parameters passed at the top level.\label{fig:graph}](media/model_graph.pdf)
 
@@ -118,11 +116,11 @@ convenient user interface. This design encourages modular development and is
 supportive of users who wish to expand functionality at different levels. The
 `caustics` package [@Stone2024] implements this code design to great effect.
 \autoref{fig:graph} shows an example `caskade` graph[^1] from `caustics`. In
-this graph the redshift parameters of each lens are linked to ensure consistent
-evaluation despite the functional backed having no explicit enforcement of this.
-See also that all of the lens objects (`ExternalShear`, `SIE`, and
-`SinglePlane`) point to a single cosmology `Module` and so share the same
-cosmological parameters automatically.
+this graph the redshift parameters (`z_l` and `z_s`) of each lens are linked to
+ensure consistent evaluation despite the functional backed having no explicit
+enforcement of this. See also that all of the lens objects (`ExternalShear`,
+`SIE`, and `SinglePlane`) point to a single cosmology `Module` and so share the
+same cosmological parameters automatically.
 
 [^1]: visual generated by `graphviz` [@graphviz]
 
@@ -145,8 +143,7 @@ Finally, `PathSim` also shares the `caskade` modular simulator building
 framework, though it focuses exclusively on time-domain dynamical systems.
 Clearly, many fields of research and development desire such modular
 simulation-building frameworks; `caskade` fulfills the role very generally,
-though not so abstractly as to require overhead schema or meta-data files. It is
-lean and extensible so as to grow with a project.
+though not so abstractly as to require overhead schema or meta-data files.
 
 
 # Acknowledgements
