### 2025-08-19 #8786 @ConnorStoneAstro
Hi @aslan-ng thanks for the positive and helpful review! I have made changes based on your comments which should be visible in the new PDF draft. I also have a PR in place for the updates from the review, which I will keep updating and then merge when the review is complete. You can see the PR here: https://github.com/ConnorStoneAstro/caskade/pull/59

The only comment that I didn't address directly is number 2. I appreciate you saying that the logo looks polished, but I still feel there are some tweaks in its future. For example I would like to add some colour and maybe downsize it. Since it's still in flux I don't want to immortalize it in the paper quite yet. If that makes sense to you.
