### 2025-08-19 ConnorStoneAstro/caskade#60 @avapolzin
[JOSS Review] Installation, contributing, tests
Hi, @ConnorStoneAstro,

The package looks like it will be very useful. Here are some comments on the installation/documentation for `caskade` as part of the JOSS Review ([openjournals/joss-reviews/#8786](https://github.com/openjournals/joss-reviews/issues/8786#issuecomment-3201663208)).

**Installation (from instructions):**
When I attempted to install `caskade` in a clean environment following the instructions in the README and RTD, I was met with a `PyTorch` error. This was easily resolved by downgrading my Python to 3.11, as PyTorch is apparently incompatible with the newest version(s?) of Python. It would be good to specify this fix somewhere in the documentation; alternatively you may want to mandate Python 3.9-3.11.

**Contributing:**
I see that you have a contributor guide in the docs, and I think this is perfectly sufficient. You may want, however, to also take advantage of GitHub’s option to add built-in “Contributing” tab alongside the README and License. This can be added by simply creating a .github/CONTRIBUTING.md file in your repository. (I consider this optional, though, so it’s not a pressing part of the review.)

**Tests (and additional installation considerations):**
- I see that tests are automated for the repository, but I did run them locally as an additional check. If people go this step, they will need to have `h5py` and `graphviz` installed, otherwise there are tests that result in errors. Potentially worth mentioning along with instructions to run local tests, as the other tests proved useful in uncovering installation errors.
- Similarly, for my computer (M1 Mac), at least, installing `jax` with `caskade` via `pip install caskade[jax]` results in an error about incompatibilty:
```bash
RuntimeError: This version of jaxlib was built using AVX instructions, which your CPU and/or operating system do not support. You may be able work around this issue by building jaxlib from source.
```
This error comes up in tests/test_backend.py (and presumably would also come up if I had tried an import before the test). There are several suggested fixes [here](https://stackoverflow.com/questions/68327863/importing-jax-fails-on-mac-with-m1-chip) including a conda-forge installation or downgrading `jax`, and the test passed after I re-installed `jax` with conda-forge. (I ultimately had to specify `jaxlib==0.4.35` in re-installing to satisfy apparently conflicting version requirements.) Discussion of the error/fix should be added to the documentation for Apple Silicon users (and possibly others).
- The local tests also demonstrated that there is a conflict between `PyTorch` and the version of `numpy` that was installed. I downgraded to `numpy==1.26.4` per the suggestion [here](https://stackoverflow.com/questions/71689095/how-to-solve-the-pytorch-runtimeerror-numpy-is-not-available-without-upgrading). This needs to be addressed in the documentation/installation instructions at minimum, and may also warrant changing requirements.txt to reflect that there are version requirements.

I will continue looking at the code functionality and open additional issues to address other points if need be.
