### 2025-08-19 9a3cfc2
Improve dependency compatibility avapolzin comments
