+++ requirements.txt
@@ -1,2 +1,2 @@
-torch
-numpy
+torch>=2.0.0
+numpy==1.26.4
