### 2025-08-19 f2ecb10
broader requirement ranges
