+++ requirements.txt
@@ -1,2 +1,2 @@
-torch>=2.0.0
-numpy==1.26.4
+torch>=2,<3
+numpy>1.22,<2
