### 2025-08-19 ConnorStoneAstro/caskade#60 @ConnorStoneAstro
Hi @avapolzin, thanks for the detailed run through of our installation/testing system! I have made a number of changes over at PR #59 . Generally, I try not to add too many strict requirements since caskade is intended to run as backend to other packages, but you are right that there are some direct incompatibilities.

* I've added a note in the installation instructions that PyTorch doesn't work with 3.12 on all systems (it worked on mine which is a linux system, but I haven't tested Mac or windows for 3.12 yet).
* I've added a contributing guide to the .github
* I updated the guide to indicate that developers should install with `pip install -e ".[dev]"` which will get the `h5py` and `graphviz` packages as well
* I've added a note in the jax install section about the jaxlib incompatibility for mac users. I'm surprised this didn't come up in tests, but maybe I just never stressed things enough for the incompatibility to cause issues
* I also added some version requirements for numpy in the requirements.txt file so that it will be less than numpy 2. Fingers crossed PyTorch will be compatible soon
