### 2025-08-19 ConnorStoneAstro/caskade#60 @avapolzin
Great; thank you, @ConnorStoneAstro! Your changes noted here adequately address the comments in this issue, so I'll close it.
