### 2025-08-20 e2a704c
update docs with list of cool features
