### 2025-08-20 #8786 @ConnorStoneAstro
@avapolzin thanks! I like the way you think, I've added a list of features as part of the front page. I'll probably periodically update this.
