### 2025-08-25 6c065de
Reviewcomments (#59)
* address asian-ng review comment 1

* aslan-ng comment 5

* Improve dependency compatibility avapolzin comments

* broader requirement ranges

* update docs with list of cool features
