### 2025-09-01 #8786 @drvinceknight
No this looks like it needs me to do a few things next but we're not far :)

Thank you @avapolzin and @aslan-ng!
