### 2025-09-11 5608932
Add active state cache decorator (#61)
