### 2025-09-15 #8786 @danielskatz
@ConnorStoneAstro - the zenodo archive doesn't seem to have all the authors of the paper. You should be able to edit the metadata of the archive to add them, without having to create a new deposit.
