### 2025-09-15 f5ac1f3
Update paper.md
