### 2025-09-15 #8786 @danielskatz
In addition, I've suggested a bunch of small changes in https://github.com/ConnorStoneAstro/caskade/pull/62 - please merge this, or let me know what you disagree with.
