### 2025-09-15 b7a5ee9
Fix typo in paper.md regarding 'encodes'
