### 2025-09-15 3d047ab
minor changes for JOSS publication (#62)
* Update paper.md

* Update paper.bib

* Fix typo in paper.md regarding 'encodes'

---------

Co-authored-by: Connor Stone, PhD <connorstone628@gmail.com>
