+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Stone2024,
-    title = {Caustics: A Python Package for Accelerated Strong Gravitational Lensing Simulations},
+    title = {Caustics: A {P}ython Package for Accelerated Strong Gravitational Lensing Simulations},
     author = {Connor Stone and Alexandre Adam and Adam Coogan and M. J. Yantovski-Barth and Andreas Filipp and Landung Setiawan and Cordero Core and Ronan Legin and Charles Wilson and Gabriel Missael Barco and Yashar Hezaveh and Laurence Perreault-Levasseur},
     year = {2024},
     journal = {Journal of Open Source Software},
@@ -74,7 +74,7 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
 
 @ARTICLE{emcee,
        author = {{Foreman-Mackey}, Daniel and {Hogg}, David W. and {Lang}, Dustin and {Goodman}, Jonathan},
-        title = "{emcee: The MCMC Hammer}",
+        title = {{emcee}: The MCMC Hammer},
       journal = {Publications of the Astronomical Society of the Pacific},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Statistics - Computation},
          year = 2013,
@@ -165,7 +165,7 @@ volume = {7},
 number = {69}, 
 pages = {3882}, 
 author = {Stammler, Sebastian M. and Birnstiel, Tilman}, 
-title = {Simframe: A Python Framework for Scientific Simulations}, 
+title = {Simframe: A {P}ython Framework for Scientific Simulations}, 
 journal = {Journal of Open Source Software} 
 } 
 
