### 2025-09-15 #8786 @danielskatz
@ConnorStoneAstro - please check this version and let me know if it looks ok
