### 2025-09-15 #8786 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
