### 2025-09-15 #8786 @danielskatz
I now see the paper at https://www.theoj.org/joss-papers/joss.08786/10.21105.joss.08786.pdf and in https://joss.theoj.org/papers/10.21105/joss.08786 (at least when I open it in a new browser).
